# Conflicts:
#	src/battle_util.c

+++ include/test/battle.h
@@ -873,7 +873,7 @@ struct HPEventContext
 
 struct StatusEventContext
 {
-    u8 status1;
+    u16 status1;
     bool8 none:1;
     bool8 sleep:1;
     bool8 poison:1;
